refactor(User) : 이름 필드에 VO(UserName) 적용 및 불변성 강화
User 엔티티의 name 필드를 String에서 UserName VO로 변경하였고,
@Embedded 및 @AttributeOverride를 통해 컬럼 매핑을 적용했습니다.
의미 있는 타입 분리를 통해 유효성 검증과 도메인 책임 분리를 의도했습니다.
