+++ src/main/java/org/terning/domain/Scraps.java
@@ -13,6 +13,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.terning.domain.common.BaseEntity;
+import org.terning.user.User;
 
 @Entity
 @Getter
